requirements.txt
browser-use==0.1.40
pyperclip==1.9.0
gradio==5.23.1
json-repair==0.40.0
langchain>=0.1.0
langchain-mistralai==0.2.4
langchain-google-genai==2.0.8
MainContentExtractor==0.0.4
python-dotenv>=1.0.0
typing-extensions>=4.8.0
requests>=2.31.0
playwright>=1.40.0
